refactor: move ToolRunner to separate file, move artifacts handlers to separate file
chore: add license
